refactor ritual music helpers
