+++ src/audio/emotion_params.py
@@ -1,0 +1,43 @@
+from __future__ import annotations
+
+"""Emotion to music parameter resolution helpers."""
+
+from pathlib import Path
+from typing import List, Tuple
+
+from INANNA_AI import emotion_analysis, sonic_emotion_mapper
+from MUSIC_FOUNDATION.inanna_music_COMPOSER_ai import (
+    SCALE_MELODIES,
+    load_emotion_music_map,
+    select_music_params,
+)
+
+EMOTION_MAP = Path(__file__).resolve().parent / "emotion_music_map.yaml"
+
+
+def resolve(emotion: str, archetype: str | None = None) -> Tuple[float, List[str], str, str]:
+    """Return tempo, melody, wave type and resolved archetype."""
+
+    if archetype is None:
+        try:  # pragma: no cover
+            archetype = emotion_analysis.get_current_archetype()
+        except Exception:  # pragma: no cover
+            archetype = "Everyman"
+
+    mapping = load_emotion_music_map(EMOTION_MAP)
+    params = sonic_emotion_mapper.map_emotion_to_sound(emotion, archetype)
+
+    tempo, _scale, melody, _rhythm = select_music_params(
+        emotion, mapping, params["tempo"]
+    )
+
+    if params.get("scale"):
+        melody = SCALE_MELODIES.get(params["scale"], melody)
+
+    wave_type = (
+        "square"
+        if any("guitar" in t or "trumpet" in t for t in params.get("timbre", []))
+        else "sine"
+    )
+
+    return tempo, melody, wave_type, archetype
