+++ tests/test_rag_music_oracle.py
@@ -10,7 +10,7 @@ from tests.helpers import emotion_stub
 from tests.helpers.config_stub import build_settings
 
 ROOT = Path(__file__).resolve().parents[1]
-sys.path.insert(0, str(ROOT))
+sys.path.insert(0, str(ROOT / "src"))
 
 # Stub heavy dependencies before importing the module
 sys.modules.setdefault("librosa", types.ModuleType("librosa"))
@@ -28,7 +28,14 @@ sys.modules.setdefault("config", config_mod)
 # Minimal play_ritual_music stub
 audio_pkg = types.ModuleType("audio")
 fake_play = types.ModuleType("play_ritual_music")
-fake_play.compose_ritual_music = lambda *a, **k: Path("out.wav")
+
+
+class _Track:
+    def __init__(self, path: Path):
+        self.path = path
+
+
+fake_play.compose_ritual_music = lambda *a, **k: _Track(Path("out.wav"))
 audio_pkg.play_ritual_music = fake_play
 sys.modules.setdefault("audio", audio_pkg)
 sys.modules.setdefault("audio.play_ritual_music", fake_play)
@@ -57,7 +64,7 @@ def test_answer_with_audio(tmp_path, monkeypatch):
     monkeypatch.setattr(
         rmo.play_ritual_music,
         "compose_ritual_music",
-        lambda e, r, **k: tmp_path / "out.wav",
+        lambda e, r, **k: _Track(tmp_path / "out.wav"),
     )
 
     text, out = rmo.answer("How does this MP3 express grief?", audio, play=True)
